+++ packages/adminui-pub/_src/app/ux/UBTinyMCETextArea.js
@@ -1,4 +1,5 @@
 require('../../ux/form/TinyMCETextArea')
+/* global tinymce */
 /**
  * Файл: UB.ux.UBTinyMCETextArea.js
  * Автор: Игорь Ноженко
@@ -9,87 +10,115 @@ Ext.define('UB.ux.UBTinyMCETextArea', {
   extend: 'Ext.ux.form.TinyMCETextArea',
   alias: 'widget.ubtinymcetextarea',
 
-  // requires: ['Ext.ux.form.TinyMCETextArea'],
-
   initComponent: function () {
-    var me = this
-    if (me.tinyMCEConfig) {
-      me.userSetup = me.tinyMCEConfig.setup
+    if (this.tinyMCEConfig) {
+      this.userSetup = this.tinyMCEConfig.setup
     }
     this.ensureTinyMCELoaded()
-    //tinyMCE.baseURL = $App.connection.baseURL + 'clientRequire/tinymce/'
-    me.tinyMCEConfig = Ext.apply({
-          // language: UB.core.UBApp.getUiLanguage(),
+    // tinyMCE.baseURL = $App.connection.baseURL + 'clientRequire/tinymce/'
+    this.tinyMCEConfig = Ext.apply({
       language_url: $App.connection.baseURL + 'models/adminui-pub/locale/tinymce/' + $App.connection.userLang() + '.js',
-      //skin_url: $App.connection.baseURL + 'clientRequire/tinymce/skins/lightgray/',
+      // skin_url: $App.connection.baseURL + 'clientRequire/tinymce/skins/lightgray/',
       table_default_attributes: {
         cellpadding: '3px',
         cellspacing: '0',
         border: '1px',
-        width: me.pageWidth && me.pageWidth > 20 ? me.pageWidth - 20 : 20,
-        style: { wordBreak: 'break-all'}
+        width: this.pageWidth && this.pageWidth > 20 ? this.pageWidth - 20 : 20,
+        style: {wordBreak: 'break-all'}
       },
       browser_spellcheck: true,
       toolbar1: 'undo redo | bold italic underline | alignleft aligncenter alignright alignjustify | styleselect formatselect fontselect fontsizeselect | print',
       plugins: [
-                // "autosave layer noneditable",
-                // disabled - " media"
         'advlist autolink lists charmap print preview hr anchor pagebreak', // link image
         'searchreplace wordcount visualblocks visualchars code ', // fullscreen
         'insertdatetime nonbreaking table contextmenu directionality', // save
-        'emoticons template textcolor image'  // paste
+        'emoticons template textcolor colorpicker image imagetools'  // paste
       ],
-      contextmenu: 'link image inserttable | cell row column deletetable',
+      contextmenu: 'link image inserttable | cell row column deletetable | rowTemplate',
       paste_data_images: true,
-
-            // content_css : "contents.css",
       statusbar: false,
       menubar: 'edit insert view format table tools',
       toolbar_items_size: 'small',
-      setup: me.onStartSetup.bind(me)
-    }, me.tinyMCEConfig)
-        // me.tinyMCEConfig.document_base_url = $App.connection.serverUrl + 'jslibs/tinymce/';
-        // me.tinyMCEConfig.relative_urls = true;
+      // image insertion
+      automatic_uploads: false,
+      // here we add custom filepicker only to Image dialog
+      file_picker_types: 'image',
+      // and here's our custom image picker
+      file_picker_callback: function (cb, value, meta) {
+        let input = document.createElement('input')
+        input.setAttribute('type', 'file')
+        input.setAttribute('accept', 'image/*')
 
-    me.tinyMCEConfig.setup = me.onStartSetup.bind(me)
+        // Note: In modern browsers input[type="file"] is functional without
+        // even adding it to the DOM, but that might not be the case in some older
+        // or quirky browsers like IE, so you might want to add it to the DOM
+        // just in case, and visually hide it. And do not forget do remove it
+        // once you do not need it anymore.
 
-        /**
-         * @event setOriginalValue
-         * @param {String} originalValue
-         * @param {Object} self
-         * fire when initialize original value
-         */
-    me.addEvents('setup', 'setOriginalValue')
+        input.onchange = function () {
+          let file = this.files[0]
+
+          // Note: Now we need to register the blob in TinyMCEs image blob
+          // registry. In the next release this part hopefully won't be
+          // necessary, as we are looking to handle it internally.
+          let reader = new window.FileReader()
+          let blob = new window.Blob([file], {type: file.type})
+          reader.addEventListener('loadend', function () {
+            let id = 'blobid' + (new Date()).getTime()
+            let blobCache = tinymce.activeEditor.editorUpload.blobCache
+            // MPV - we need a base64 representation to embed image to document
+            let b64 = reader.result.split(',', 2)[1]
+            let blobInfo = blobCache.create(id, blob, b64)
+            blobCache.add(blobInfo)
+            // call the callback and populate the Title field with the file name
+            cb(blobInfo.blobUri(), { title: file.name })
+          })
+          reader.addEventListener('error', function (event) {
+            throw new Error(event)
+          })
+          reader.readAsDataURL(blob)
+        }
+        input.click()
+      }
+    }, this.tinyMCEConfig)
+
+    this.tinyMCEConfig.setup = this.onStartSetup.bind(this)
+
+    /**
+     * @event setOriginalValue
+     * @param {String} originalValue
+     * @param {Object} self
+     * fire when initialize original value
+     */
+    this.addEvents('setup', 'setOriginalValue')
 
     this.callParent(arguments)
   },
 
   afterRender: function () {
-    var me = this
-
-    me.callParent(arguments)
-    if (!me.wysiwygIntialized) {
-      me.updateLayout()
+    this.callParent(arguments)
+    if (!this.wysiwygIntialized) {
+      this.updateLayout()
     }
   },
 
   onStartSetup: function (ed) {
-    var me = this
-        // ed.baseURI = $App.connection.serverUrl + 'jslibs/tinymce1/';
+    let me = this
+
     tinymce.baseURL = $App.connection.baseURL + 'clientRequire/tinymce/'
     this.fireEvent('setup', ed)
     if (this.userSetup) {
       this.userSetup(ed)
     }
-        // todo Update layout when frame ready. Perhaps there is a better solution.
-    ed.on('init', function (e) {
+    // todo Update layout when frame ready. Perhaps there is a better solution.
+    ed.on('init', function () {
       if (me.ownerLayout) {
         me.ownerLayout.onContentChange()
       }
     })
   },
 
-    // for get focus in BasePanel
+  // for get focus in BasePanel
   isFocusableField: true,
   isFocusable: function () {
     return !this.readOnly
@@ -105,14 +134,12 @@ Ext.define('UB.ux.UBTinyMCETextArea', {
    * @returns {Promise}
    */
   setSrc: function (cfg) {
-    var me = this
-    var blobData = cfg.blobData
-    var resetOriginalValue = cfg.resetOriginalValue
+    let me = this
 
     function onDataReady (response) {
       me.suspendCheckChange = true
       me.setValue(response)
-      if (resetOriginalValue) {
+      if (cfg.resetOriginalValue) {
         me.resetOriginalValue()
         me.fireEvent('setOriginalValue', response, me)
       }
@@ -120,9 +147,10 @@ Ext.define('UB.ux.UBTinyMCETextArea', {
       return response
     }
 
+    let blobData = cfg.blobData
     if (blobData) {
       return new Promise((resolve, reject) => {
-        var reader = new FileReader()
+        let reader = new window.FileReader()
         reader.addEventListener('loadend', function () {
           resolve(onDataReady(reader.result))
         })