+++ packages/adminui-pub/CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [4.2.12]
+### Added
+- AdminUI: In case of first login attempt LoginWindow will activate a tab for a first auth method from server config `security.authenticationMethods` array
+
+### Fixed
+
+
+## [4.2.11]
 ### Added
 - UBBoxSelect now accept ubRequest as a config parameter (for store creation)
+- UBReportEditor can insert image from file (Insert -> Image -> click on button with photo)
+- UBReportEditor added build-in image editor (click on image to actiate)
 
 ### Fixed
 - since form definitian now evaluated only once (HMR) both `EntityGridPanel` & `ubdetailgrid` now accept `customActionas` as a Ext.Action config (not a class instance)